fix: :wrench: file existence validation for local creation

+++ src/commands/service/local_service.py
@@ -22,7 +22,20 @@ class LocalFileService:
         # crate folders (even if they exist or not)
         makedirs(dir_path, exist_ok=True)
 
-        # create file
         file_path = path.join(dir_path, name)
+
+        # if file exists return
+        if path.exists(file_path):
+            return {
+                'ok': False,
+                'msg': f"El archivo '{name}' ya existe en la ruta '{relative_path}'"
+            }
+
+        # create file
         with open(file_path, 'w') as file:
             file.write(body)
+
+        return {
+            'ok': True,
+            'msg': f"Archivo '{name}' creado con exito en la ruta '{relative_path}'"
+        }
